Move demo to SDL2
