+++ nk/impl_sdl_gl2.go
@@ -8,7 +8,7 @@ package nk
 #include <string.h>
 
 #define NK_IMPLEMENTATION
-#define NK_GLFW_GL2_IMPLEMENTATION
+#define NK_SDL_GL2_IMPLEMENTATION
 
 #include "nuklear.h"
 */
