+++ nk/impl_sdl_gl3.go
@@ -9,7 +9,7 @@ package nk
 #include <string.h>
 
 #define NK_IMPLEMENTATION
-#define NK_GLFW_GL3_IMPLEMENTATION
+#define NK_SDL_GL3_IMPLEMENTATION
 
 #include "nuklear.h"
 */
